Change the type of return to Long and add the index validation.

+++ src/test/java/com/dev/dro/AlticciServiceTest.java
@@ -1,6 +1,6 @@
 package com.dev.dro;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import io.quarkus.test.junit.QuarkusTest;
 import javax.inject.Inject;
@@ -14,25 +14,25 @@ class AlticciServiceTest {
 
   @Test
   void getAlticciByIndexZero() {
-    Integer result = service.getAlticciByIndex(0);
+    Long result = service.getAlticciByIndex(0);
     assertEquals(0, result);
   }
 
   @Test
   void getAlticciByIndexOne() {
-    Integer result = service.getAlticciByIndex(1);
+    Long result = service.getAlticciByIndex(1);
     assertEquals(1, result);
   }
 
   @Test
   void getAlticciByIndexTwo() {
-    Integer result = service.getAlticciByIndex(2);
+    Long result = service.getAlticciByIndex(2);
     assertEquals(1, result);
   }
 
   @Test
   void getAlticciByIndexEight() {
-    Integer result = service.getAlticciByIndex(8);
+    Long result = service.getAlticciByIndex(8);
     assertEquals(5, result);
   }
 }
